renderer: Redesign of video_pipeline
Design was changed so that multiple sources are rendered into a
final texture, which can be draw to the screen multiple times
using a normal draw command.

The reason behind the change, is that we need to be able to read out
the final image so we can send it to the video encoder. Otherwise
the work of combining images has to be done in software anyways