core: retroactively assign source information to source providers that don't support querying
In Pipewire for example, we can't query information about monitors available on the
system that the user may select for screen sharing. So instead of setting this
information in the model when the screencapture backend initializes, we do it
one by one when streams from that backend are opened. Since we still don't have names
for the streams, we just assign a generic one