renderer: fix validation error when resizing render_pass
Previously we held the number of pixels allocated for multisample
and depth buffer, and only re-allocated when we needed to grow the
swapchain of a larger dimension.

Now, we use the provided function to query the memory requirements
of the images when re-create them and decide whether a re-allocation
is required from that.

This fixes validation error that triggers from binding an image to
a memory handle that is too small